Person Assessment function

Estimate Person Paramters and calculate Person Fit in one step to gain resonse pattern assessment. Submit a data.frame which contains item responses, or an fitted model (Rasch Model and Partial Credit Model are supported) of the eRm package.

PPass(...)

# S3 method for default
PPass(respdf, items = "all", mod = c("1PL", "2PL",
  "3PL", "4PL", "PCM", "GPCM", "MIXED"), fitindices = c("lz", "lzstar",
  "infit", "outfit"), ...)

# S3 method for Rm
PPass(RMobj, fitindices = c("lz", "lzstar", "infit",
  "outfit"), ...)

Arguments

respdf
A data.frame which contains the items, and perhaps other informations. Each row is a person related resonse patter. Each column denotes a variable.

items
A numeric (integer) vector which indicates the positions of the items in the data.frame (respdf). If items = 'all', all columns are treated as items.

mod
Choose your data generating model. This argument switches between the three person parameter estimating functions PP_4pl, PP_gpcm and PPall.

fitindices
A character vector which denotes the fit indices to compute.

RMobj
A fitted Rasch Model (RM()) object which stems from the eRm package.

…
Submit arguments to the underlying functions: PP_4pl, PP_gpcm and PPall (see documentation files) for person parameter estimation.

Value

The original data.frame and

The Person Parameter estimates incl. Standard Errors (2 columns)

Person Fit Indices you chose (1 or more)

Details

PPass fuses Person Parameter estimation and Person Fit computation into a single function.

See also

Examples

library(eRm)
### real data ##########

data(pp_amt)

d <- pp_amt$daten_amt

rd_res <- PPass(respdf = d,
                items = 8:ncol(d),
                mod="1PL",
                thres = pp_amt$betas[,2],
                fitindices = "lz")

#> Estimating:  1pl model ... 
#> type = wle 
#> Estimation finished!

head(rd_res)

#>    estimate        SE        lz   lz_unst
#> 1 1.4098572 0.4177670 -0.631024 -0.613445
#> 2 0.3515596 0.3728684  1.229632 -0.627897
#> 3 1.0745004 0.3741313  1.097723 -0.627596
#> 4 0.3204585 0.6900186  0.981576 -0.494311
#> 5 1.3879159 0.4643975 -0.203399 -0.600039
#> 6 0.9457363 0.4184999  0.573956 -0.615722

## ========== RM - eRm 
my_data <- eRm::sim.rasch(200, 12)
my_rm <- eRm::RM(my_data)

res_pp1 <- PPass(my_rm)

#> Estimating:  1pl model ... 
#> type = wle 
#> Estimation finished!

## ========== PCM - eRm

set.seed(2751)

THRES  <- matrix(c(-2,-1.23,1.11,3.48,1
                   ,2,-1,-0.2,0.5,1.3,-0.8,1.5),nrow=2)
THRES <- rbind(THRES,c(-0.2,NA,NA,NA,NA,NA))

sl          <- rep(1,6)
THRESx      <- rbind(0,THRES)
THETA       <- rnorm(200)
simdat_gpcm <- sim_gpcm(thres = THRESx,alpha = sl,theta = THETA)

my_pcm <- eRm::PCM(simdat_gpcm)

res_pp2 <- PPass(my_pcm)

#> Estimating: GPCM ... 
#> type = wle 
#> Estimation finished!

#> Warning: Only 'infit and outfit' are currently supported. The calculation is executed with infit outfit 

## ==========  1PL model

set.seed(1337)

# intercepts
diffpar <- seq(-3,3,length=15)
# slope parameters
sl     <- round(runif(15,0.5,1.5),2)
la     <- round(runif(15,0,0.25),2)
ua     <- round(runif(15,0.8,1),2)

# response matrix
awm <- matrix(sample(0:1,100*15,replace=TRUE),ncol=15)
awm <- as.data.frame(awm)
# estimate person parameter

# estimate person parameter and person fit
out <- PPass(respdf = awm,thres = diffpar, items="all",
             mod=c("1PL"), fitindices= c("lz","lzstar","infit","outfit"))

#> Estimating:  1pl model ... 
#> type = wle 
#> Estimation finished!

# show first rows
head(out)

#>     estimate        SE        lz   lz_unst    lzstar    infit     in_t in_chisq
#> 1 -0.2283002 0.6819765 -6.409070 -1.327819 -6.414996 3.139671 3.958653   76.751
#> 2 -0.6878210 0.6874864 -7.163429 -1.411064 -7.228344 3.295823 4.169337  107.918
#> 3  0.2283002 0.6819765 -6.201196 -1.299247 -6.206937 2.979662 3.749448   86.037
#> 4 -0.6878210 0.6874864 -5.478936 -1.182492 -5.529163 2.793206 3.507021   84.946
#> 5 -0.6878210 0.6874864 -6.531744 -1.325349 -6.591151 3.007458 3.798334  107.702
#> 6  1.1569232 0.6998949 -1.811795 -0.662741 -1.872497 1.789309 1.902399   22.509
#>   in_df in_pv   outfit     ou_t ou_chisq ou_df ou_pv
#> 1    14 0.000 5.116709 3.676194   76.751    14 0.000
#> 2    14 0.000 7.194532 4.169166  107.918    14 0.000
#> 3    14 0.000 5.735802 3.996835   86.037    14 0.000
#> 4    14 0.000 5.663040 3.543487   84.946    14 0.000
#> 5    14 0.000 7.180157 4.163728  107.702    14 0.000
#> 6    14 0.069 1.500628 0.789110   22.509    14 0.069